[web] Move CanvasKit fragment shader classes to canvaskit/fragment_shader.dart
Moves CkFragmentProgram, CkFragmentShader, and uniform slot classes from canvaskit/painting.dart to canvaskit/fragment_shader.dart to decouple fragment shader compilation from paint descriptor logic in preparation for ui.Paint unification.
